requirements.txt
attrs==23.1.0
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
Flask==3.0.0
Flask-Cors==4.0.0
greenlet==3.0.0
gunicorn==21.2.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mysqlclient==2.2.0
packaging==23.2
password-validator==1.0
PyJWT==2.8.0
referencing==0.30.2
requests==2.31.0
rpds-py==0.10.3
schedule==1.2.1
SQLAlchemy==2.0.21
typing_extensions==4.8.0
urllib3==2.0.6
uuid==1.30
validator-collection==1.5.0
Werkzeug==3.0.0
